+++ docs/changes.rst
@@ -9,6 +9,8 @@ Not yet released.
 
 **Bug fixes**
 
+* Formatting of some :ref:`audit-log` entries.
+
 **Compatibility**
 
 **Upgrading**